requirements.txt
streamlit
python-pptx
python-docx
Pillow
pdfplumber
requests
matplotlib
anthropic
python-dotenv
PyPDF2
groq
